+++ libs/langchain/langchain_classic/memory/buffer.py
@@ -5,7 +5,7 @@ from langchain_core.messages import BaseMessage, get_buffer_string
 from langchain_core.utils import pre_init
 from typing_extensions import override
 
-from langchain_classic.memory.base import BaseMemory
+from langchain_classic.base_memory import BaseMemory
 from langchain_classic.memory.chat_memory import BaseChatMemory
 from langchain_classic.memory.utils import get_prompt_input_key
