+++ libs/langchain/langchain_classic/memory/chat_memory.py
@@ -10,7 +10,7 @@ from langchain_core.chat_history import (
 from langchain_core.messages import AIMessage, HumanMessage
 from pydantic import Field
 
-from langchain_classic.memory.base import BaseMemory
+from langchain_classic.base_memory import BaseMemory
 from langchain_classic.memory.utils import get_prompt_input_key
