+++ libs/langchain/langchain_classic/memory/vectorstore.py
@@ -8,7 +8,7 @@ from langchain_core.documents import Document
 from langchain_core.vectorstores import VectorStoreRetriever
 from pydantic import Field
 
-from langchain_classic.memory.base import BaseMemory
+from langchain_classic.base_memory import BaseMemory
 from langchain_classic.memory.utils import get_prompt_input_key
