+++ libs/langchain/langchain_classic/schema/memory.py
@@ -1,3 +1,3 @@
-from langchain_classic.memory.base import BaseMemory
+from langchain_classic.base_memory import BaseMemory
 
 __all__ = ["BaseMemory"]
